Create Lunch entity and tests
A Lunch represents the combination of a recipe, and the actual ingredients
that can be used to make the recipe.

The Lunch can be created from a passed in Recipe entity, and a Doctrine
ArrayCollection of Ingredient entities.

It can tell us whether it contains ingredients which match a passed in list of
names.

It can tell us what its least fresh ingredient is.

Require the Doctrine Collections package, to make working with arrays of
Ingredients easier and cleaner.