Create FileSystemDataLoader and tests
A FileSystemDataLoader will be able to load recipes and ingredients from JSON
formatted files in the filesystem, and return them as ArrayCollections.

It will check the file exists, and that the JSON data in the file is valid.

Add the vfsStream package, which mocks a fileystem so the tests don't have to
rely on the environment's file system to be available.