Create Fridge Entity and tests
A Fridge can contains recipes and ingredients.

It can load ingredients and recipes from different data sources.
It can assemble lunches from the provided recipes and ingredients.
For given dates, it provides fresh ingredients, fresh lunches, and recipes
that can be made using the fresh ingredients in the Fridge, or from a list of
provided ingredients.

It won't provide any recipes or ingredients which are past their use-by date.

When a recipe requires ingredients which might be past their best by date, it
will sort those recipes to the bottom of the list of recipes returned.