Add Lunch API controller, routes, and tests
Instantiate a Fridge, load it with ingredients, recipes, and assemble lunches
from the available ingredients and recipes.

Return recipes which use ingredients that are within their use-by dates, with
the recipes requiring the ingredients that are past their best-before dates
sorted to the bottom of the response.

Accept an optional date query string parameter, and fallback to the date of the
request if no date provided.

Respond with an error object and the appropriate HTTP response code if there is
a problem with the format or value of the date query string parameter.

Typically, Symfony best practice is to use annotations above the methods in the
controller for the routing.
In this instance, I've chosen to go with adding a single route in the
config.routes file, because it requires less setup, and when I tried
to install the symfony annotations package, there were conflicts with the
sensio-extra package versions which would have required other packages to
be upgraded.
At this time I felt it was appropriate to maintain the version of symfony which
was provided in the original codebase, as sometimes we don't always have the
option of being able to upgrade at a given time, and need to find other
solutions.