Dockerise project with test support
Create images for the PHP and NGINX containers, built using Docker Compose.
Update the configuration of the Symfony app to support running tests in the
Docker project.
Create a Makefile to simplify the build, up, test, and down commands for the
project.
Update README with instructions for running in Docker.